move helpers into structure folder
