+++ src/packages/core/content-type/index.ts
@@ -1,6 +1,4 @@
 export * from './components/index.js';
-export * from './content-type-container-structure-helper.class.js';
-export * from './content-type-property-structure-helper.class.js';
-export * from './content-type-structure-manager.class.js';
 export * from './repository/index.js';
+export * from './structure/index.js';
 export * from './types.js';
