+++ src/packages/core/content-type/structure/content-type-property-structure-helper.class.ts
@@ -1,5 +1,5 @@
+import type { UmbContentTypeModel, UmbPropertyContainerTypes, UmbPropertyTypeModel } from '../types.js';
 import type { UmbContentTypePropertyStructureManager } from './content-type-structure-manager.class.js';
-import type { UmbContentTypeModel, UmbPropertyContainerTypes, UmbPropertyTypeModel } from './types.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbArrayState, UmbObserverController } from '@umbraco-cms/backoffice/observable-api';
