+++ src/packages/core/content-type/structure/index.ts
@@ -1,0 +1,3 @@
+export * from './content-type-container-structure-helper.class.js';
+export * from './content-type-property-structure-helper.class.js';
+export * from './content-type-structure-manager.class.js';
